Bind the iOS UI the presenters

+++ sample/shared/src/commonMain/kotlin/io/github/andremion/boomerang/sample/data/TaskDataSource.kt
@@ -6,7 +6,7 @@ suspend fun getTasks(): List<Task> {
     delay(500) // Simulate long running operation
     return List(30) { index ->
         Task(
-            id = index,
+            id = index.toLong(),
             name = "Task # $index",
             isDone = false
         )
